♻️ refactor : 검색, 로그인 페이지 시안과 같게 만들고 리팩토링. input border 없앰.
