+++ src/app/login/_components/ErrorMessage.tsx
@@ -1,0 +1,5 @@
+const ErrorMessage = ({ children }: { children: React.ReactNode }) => {
+  return <span className="text-[12px] text-[#FC828A]">{children}</span>;
+};
+
+export default ErrorMessage;
