+++ src/app/search/[id]/_components/ResponseContent.tsx
