+++ src/app/search/[id]/_components/SearchResultCount.tsx
